+++ tests/Adapter/QuoteHandlerTest.php
@@ -14,18 +14,16 @@ class QuoteHandlerTest extends TestCase
 
     protected function setUp(): void
     {
-        $this->handler = new QuoteHandler(function ($value) {
+        $this->handler = new QuoteHandler(function ($value): string {
             return "`{$value}`";
         });
         parent::setUp();
     }
 
     /**
-     * @param mixed $value
-     * @param mixed $expected
      * @dataProvider valueDataProvider
      */
-    public function testValue($value, $expected): void
+    public function testValue(mixed $value, string $expected): void
     {
         static::assertSame($expected, $this->handler->value($value));
     }
@@ -55,10 +53,9 @@ class QuoteHandlerTest extends TestCase
     }
 
     /**
-     * @param mixed $value
      * @dataProvider intoDataProvider
      */
-    public function testInto(string $expected, string $text, $value): void
+    public function testInto(string $expected, string $text, mixed $value): void
     {
         static::assertSame($expected, $this->handler->into($text, $value));
     }
@@ -77,10 +74,9 @@ class QuoteHandlerTest extends TestCase
     }
 
     /**
-     * @param string|string[] $ident
      * @dataProvider columnAsData
      */
-    public function testColumnAs(string $expected, $ident, string $alias, ?bool $auto): void
+    public function testColumnAs(string $expected, string|array $ident, string $alias, ?bool $auto): void
     {
         $result = ($auto !== null) ?
             $this->handler->columnAs($ident, $alias, $auto) :
@@ -99,10 +95,9 @@ class QuoteHandlerTest extends TestCase
     }
 
     /**
-     * @param string|string[] $ident
      * @dataProvider tableAsData
      */
-    public function testTableAs(string $expected, $ident, string $alias, ?bool $auto): void
+    public function testTableAs(string $expected, string|array $ident, string $alias, ?bool $auto): void
     {
         $result = ($auto !== null) ?
             $this->handler->tableAs($ident, $alias, $auto) :
@@ -120,10 +115,9 @@ class QuoteHandlerTest extends TestCase
     }
 
     /**
-     * @param string|string[] $ident
      * @dataProvider identifierData
      */
-    public function testIdentifier(string $expected, $ident, ?bool $auto): void
+    public function testIdentifier(string $expected, string|array|SqlFragment $ident, ?bool $auto): void
     {
         $result = ($auto !== null) ?
             $this->handler->identifier($ident, $auto) :