+++ tests/Exception/PDOExceptionStub.php
@@ -9,7 +9,7 @@ class PDOExceptionStub extends \PDOException
     /**
      * @param int|string $code
      */
-    public function __construct(string $message, $code, \Exception $previous = null)
+    public function __construct(string $message, $code, ?\Exception $previous = null)
     {
         parent::__construct($message, 0, $previous);